[test] Move assertMetadataEqual to AbstractInstallerTest
